Properly unmarshal field values.
The indirection of the concrete type (e.g. `[]CategoryValue`) through
the `interface{}` in `Field.Values` makes the reflection under
`json.Unmarshal` assign `Field.Values` to a generic
`map[string]interface{}`.

By passing in the concrete type and assigning after we avoid that. It is
however quite a bit more verbose and repetitive.